requirements.txt
# requirements.txt
phidata
python-dotenv